move camera into seperate directory
