+++ field_friend/vision/__init__.py
@@ -5,13 +5,9 @@ from .camera_configurations import configurations
 from .camera_configurator import CameraConfigurator
 from .simulated_cam import SimulatedCam
 from .simulated_cam_provider import SimulatedCamProvider
-from .zedxmini_camera import ZedxminiCamera
-from .zedxmini_camera_provider import ZedxminiCameraProvider
 
 __all__ = [
     'CalibratableUsbCamera',
-    'ZedxminiCamera',
-    'ZedxminiCameraProvider',
     'CalibratableUsbCameraProvider',
     'SimulatedCam',
     'SimulatedCamProvider',
